Create helper scripts to simplify filter commands
These changes revealed that the previous method was not generating salts
as expected, so unfortunately this commit is a backward incompatible
change. There must have been some quoting inconsistencies with
constructing the long one-liners. The salt generation for the clean
filter in particular was not outputting the expected results, and I had
not verified things by hand before. Now when I run the salt generation
manually, eveything matches what transcrypt generates.