Add ability to dump the raw commit objects for all encrypted files
Using the wildcard with the --show-raw flag will now dump all of the
encrypted files' commit objects. This is useful for debugging issues
like a failed smudge filter, which usually happens due to user error
when a file is staged before setting the *crypt* Git attribute.

To expand on that, if you see errors like:

    error: external filter "$(git rev-parse --show-toplevel)"/.git/crypt/smudge failed 1

...you can run `transcrypt --show-raw '*'` and look for any file that
is being stored as plain text, and that's your culprit.