Support both old and new mktemp versions on OS X
My previous idea didn't work because the $GIT_DIR variable is not
actually set in the script and isn't expanded during creation due to how
we cat those files (a lot of complicated escaping of other variables
would be necessary otherwise).

This conditional should be more universal, where the first `mktemp`
command should work on Linux and OS X systems 10.11 (El Capitan) and
newer. On older OS X systems, the command will fail, so we ignore the
STDERR "usage:" output, and call the command with a template.