Add support for multiple passphrase "contexts" in one repo
This is a very rough but working proof of concept
showing how transcrypt could be used to encrypt
different sets of files using different passphrases
within a single repository.

It shows that the process works fairly well, with
just those files for which the user has set the
appropriate passphrase in the appropriate "context"
being decrypted, and the rest left as they were.

Example usage:

Choose a name for a non-default encryption "context"
e.g. "super-secret"

Set `.gitattributes` as usual for default decryption
and with the "super-secret" suffix for that context:

  secret.txt filter=crypt diff=crypt
  super-secret.txt filter=crypt-super-secret diff=crypt-super-secret

Init transcrypt in default and "super-secret" contexts:

  $ transcrypt
  ... follow prompts

  $ CONTEXT=-super-secret transcrypt
  ... follow prompts

Write your secret files, `git add` them and commit.

Clone repo elsewhere, and confirm that you can or
cannot see file contents as you apply some
combination of init commands in the original repo
as given by:

  $ transcrypt --display

  $ CONTEXT=-super-secret transcrypt --display