Add `--list-contexts` option
Listing contexts is a complex business because
contexts could be defined in two places, the git
config or the .gitattributes file, and we should
list contexts defined in either place.

This command finds context names wherever they
are configured, and prints them out along with
status info about whether the context is correctly
set up in the git config or attributes file.