Clearer logic for `ls-crypt` alias
Changed the logic for context-specific ls-crypt
aliases such that:

- the pre-existing `ls-crypt` alias remains and is
  the "global" command to list all encrypted files
  regardless of context (the `ls-crypt-all` did
  this job before this change)
- for each context *including* the 'default' one
  a specific `ls-crypt` command variant is created
  to list only the files in that context, e.g:
  ls-default-crypt, ls-top-secret-crypt, etc

The upside of this change is that the existing
`ls-crypt` alias will work more as users would
expect, even as they add contexts.

The downside is that we end up with an additional
`ls-default-crypt` alias defined even if a user
doesn't add any extra contexts.

TODO: Do we even need context-specific `ls-crypt`
variants? It is only used in force_checkout()
to checkout only those files encrypted in the
current context, but do we need to be so specific
even there?