Add Git pre-commit hook to abort commit with unencrypted files
On init, add a pre-commit Git hook script to check
Transcrypt-managed files and abort a commit if
there is an un-encrypted file staged in the index
that would otherwise be committed in plaintext.

This is a safety mechanism to prevent accidental
commits of plain text files that have been staged
by tools that do not respect or run the
.gitattribute filters that Transcrypt needs to do
its job.

On commit failure, the error message says how to
re-stage the file using Git on the command line.

Because Git hooks work with single scripts only
it is difficult to cleanly install and uninstall
hook scripts, especially if the user already has
a pre-commit hook script in place.

To handle this situation cleanly if naively

- on init, check if the user already has a
  pre-commit script and instead of clobbering it,
  save the script to the working directory and
  print a message asking the user to install it
- on uninstall, rename the pre-commit script
  instead of deleting and tell the user we have
  done so, to avoid deleting a script the user
  might have customised.