Defer errexit to as late as possible
At the moment, transcrypt immediately exits with a failure when run
outside of a git repository, regardless of what command is given to
transcrypt. This causes unexpected behavior when running commands like
"help" or "version".

This commit moves errexit to as late as possible, while still keeping it
as a safe guard for the time being. Ideally, errexit would not be used
in transcrypt at all.