Fix safety checks for --list
Prior to this change `transcrypt --list` would fail in a configured
repo due to the checks in `run_safety_checks()`:

    transcrypt: the current repository is already configured; see --display

If the `--list` option should be available in un-configured repos
the checking logic will need to be made more sophisticated, or maybe
just skip the `$requires_existing_config` checks for list commands?

NOTE: this bug was found through the new tests.