Permit --list operation regardless of repo's config status
Prior logic required the current repository to either:

    require configuration && be configured

    not require configuration && not be configured

This logic missed a necessary third case: work whether or not the
repo is configured.

This case could not be expressed with the existing flags, so I added
the `ignore_config_status` flag just for --list. It might be useful
for other operations as well?