Add Git pre-commit hook to abort commit with unencrypted files
On transcrypt init, add a pre-commit Git hook script to check crypt
files and abort a commit if there is an un-encrypted file staged in
the index that would otherwise be committed in plaintext.

This is a safety mechanism to prevent accidental commits of plain
text files that have been staged by tools that do not respect or
run the .gitattribute filters that Transcrypt needs to do its job.

A pre-commit-crypt file is now always installed in the Git hooks
directory, but is only "activated" if there is no existing
pre-commit hook file.

On uninstall the pre-commit file is removed only if it is definitely
safe to do so because the file exactly matches the hook file put in
place by Transcrypt, proving the user has made no manual changes.