Fix init of repo with --force when secrets file is deleted
Init-ing a dirty repo where secrets files have been deleted should
work with the --force option, which forcibly checks out each secrets
file.

However, due to the change to use stricter bash mode in version 2,
the `rm` command preceding the force checkout can cause all of
transcrypt to fail if the `rm` triggers an error when the secrets
file is deleted from the working copy.

This changes the delete command to `rm -f` to delete the secrets
file if present, and ignore it if not.