Make path to Git crypt/ directory somewhat configurable #104
Build and use a `CRYPT_DIR` variable for the path to the crypt/
directory where transcrypt writes its scripts.

Re-use this `CRYPT_DIR` variable throughout the script to avoid the
need to rebuild the crypt/ path multiple times.

ALSO write the `CRYPT_DIR` path the new setting `transcrypt.crypt-dir`
so it can be easily re-used when a dynamic lookup of the crypt/
directory is needed, i.e. for the pre-commit script and for configs
that point to the clean, smudge, textconv, and merge commands.

In addition to centralising some path calculation, this change allows
advanced users to pre-set the `transcrypt.crypt-dir` config before
initialising transcrypt to control where the crypt/ directory is
stored (or even what it is named).