Respect but DO NOT write transcrypt.crypt-dir setting #104
Allow advanced users to override the default path to a crypt/ dir
with the optional `transcrypt.crypt-dir` setting, but do not write
this setting because it bakes an absolute path into transcrypt's
Git config which will break if the user moves the repo directory.

We assume advanced users will avoid this problem with their custom
`transcrypt.crypt-dir` settings, or be willing to fix the breakage.