Consolidate all git operation scripts into a single transcrypt script
Instead of installing separate scripts for the git clean, smudge,
textconv, and merge operations a single *transcrypt* script that
contains all these sub-scripts is installed instead.

This reduces the clutter in the repository's _crypt/_ directory and
makes it easier to develop and test *transcrypt* itself.

Also add support for an optional `transcrypt.crypt-dir` setting for
advanced users to override the path of the _.git/crypt/_ directory to
permit things like installing transcrypt in a repository on a
device without execute permissions.