Ensure encrypted files include required "Salted__" prefix even for OpenSSL 3
Recognise when a file is encrypted with a version of OpenSSL or
similar that does not include the "Salted__" prefix in the cipher
text prior to Base64 encoding, and add that prefix.

This makes Transcrypt compatible with OpenSSL version 3 which changed
the behaviour of salting files and leaves off the "Salted__" prefix.

This workaround requires another temporary file to store encrypted
data, to check for the prefix before Base64 encoding the cipher text
into Git's index.