Change approach to use ugly but tractable crypt-<context-name> git attr values instead of crypt-context attr
Unfortunately we cannot rely on a single `crypt-context` attribute
in .gitattributes to mark a file(s) path as handled by a specific
context, because the `textconv` (and it alone) does not receive the
original file name during processing so cannot derive the context
name from the `crypt-context` attr for the file being processed.

To handle all file processing use-cases I have reverted to my
original approach, where git attributes must use context-specific
values to capture the context name in all the places we need to
look it up later.

In short, the pattern in .gitattributes for context "admin" will be:

top-secret filter=crypt-admin diff=crypt-admin merge=crypt-admin