Fix unbound variable errors applying stash of secret file
The GIT_REFLOG_ACTION and THEIRS_LABEL environment variables are
normally available when doing a merge with git, but are undefined
when merging a stash.

This change and the prior one avoids the `unbound variable` error
when these variables are unset by defaulting to empty values, and
falling back to "theirs" as the in-merge description of incoming
changes from the stash.