Avoid null bytes decryption work-around for files with doubled "Salted" prefix due to #147
Improve the work-around for files incorrectly encrypted with doubled
"Salted" prefixes due to #147 to avoid null byte warnings – and
possibly errors – by checking prefix data as hex-encoded values
instead of raw bytes which could contain null characters that are not
well handled in bash scripts.

This should avoid warnings like the following during decryption:

    warning: command substitution: ignored null byte in input

Unfortunately the need for hex-encoding of bytes adds a new
requirement for the `hexdump` command.