Revert ill-judged attempt to auto-fix faulty encrypted files with doubled "Salted" prefix #158
The automatic fix for mistakenly double-salted encrypted files seems
to be causing file- and system-dependent failures for people, and is
almost certainly not worth this hassle, or the performance overhead,
since I haven't heard of anyone having a double-salted file to
recover.

Reverting this now as a bad idea, to return the `smudge` operation
back to its original implementation which is massively simpler and
hopefully no longer sometimes broken.

This reverts commits:
- f709d4718ed913bd298c08b6c1e38f7bcf3ccaa8.
- b380f2c90a08f0373001732c9f05d1218d115caa.