Merge Release 2.2.3
Revert ill-judged attempt to auto-fix faulty encrypted files with
doubled "Salted" prefix #158

The automatic fix for mistakenly double-salted encrypted files seems
to be causing file- and system-dependent failures for people, and is
almost certainly not worth this hassle, or the performance overhead,
since I haven't heard of anyone having a double-salted file to
recover.

Reverting this now as a bad idea, to return the `smudge` operation
back to its original implementation which is massively simpler and
hopefully no longer sometimes broken.